+++ app/src/main/java/org/maplibre/navigation/android/example/MainActivity.java
@@ -39,9 +39,14 @@ public class MainActivity extends AppCompatActivity implements PermissionsListen
 
         // Specify an adapter
         list.add(new SampleItem(
-                getString(R.string.title_mock_navigation),
-                getString(R.string.description_mock_navigation),
-                MockNavigationActivity.class
+            getString(R.string.title_mock_navigation),
+            getString(R.string.description_mock_navigation),
+            MockNavigationActivity.class
+        ));
+        list.add(new SampleItem(
+            getString(R.string.title_valhalla_navigation),
+            getString(R.string.description_vallhalla_navigation),
+            ValhallaNavigationActivity.class
         ));
         list.add(new SampleItem(
             getString(R.string.title_navigation_ui),