+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Maplibre welcomes participation and contributions from everyone.
   - Remove Mapbox telemetry provider references (the references to code have already been removed)
 - Additional build version checks `if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU)` for deprecated context register receiver calls.
 - Added directions builder `waypointIndices` to NavigationRoute builder.
+- Added sample code on how to use the Valhalla routing server directly in `ValhallaNavigationActivity`. Please make sure to add this line to the `app/main/res/values/developer-config.xml`:
+  ```xml
+    <string name="valhalla_url" translatable="false">https://valhalla1.openstreetmap.de/route</string>
+  ```
+- Fix crash on ManeuverView, caused by `contains` not being able to handle null values [#122](https://github.com/maplibre/maplibre-navigation-android/issues/122)
 
 ### v3.0.0 - November 5, 2023